refactor: extract shared OTLP utilities to reduce code duplication
Create lib/otlp-utils.ts with shared:
- LogAttributes type
- SERVICE_NAME, POSTHOG_API_KEY, POSTHOG_HOST constants
- getLogsEndpoint, getSeverityNumber, formatAttributeValue functions
- buildOTLPPayloadSingle and buildOTLPPayloadBatch functions

Update posthog-logger.ts and logging-middleware.ts to import from
the shared module instead of duplicating the implementation.
