+++ lib/logging-middleware.ts
@@ -5,12 +5,13 @@
  * Captures: action start, success, error, duration, and key parameters.
  */
 
-import { posthogLogger, LogAttributes } from './posthog-logger';
-
-// PostHog configuration for direct sending (edge runtime compatible)
-const POSTHOG_API_KEY = process.env.POSTHOG_API_KEY;
-const POSTHOG_HOST = process.env.POSTHOG_HOST || process.env.NEXT_PUBLIC_POSTHOG_HOST || 'https://eu.i.posthog.com';
-const SERVICE_NAME = 'mietfluss';
+import { posthogLogger } from './posthog-logger';
+import {
+    LogAttributes,
+    POSTHOG_API_KEY,
+    getLogsEndpoint,
+    buildOTLPPayloadSingle,
+} from './otlp-utils';
 
 export interface ActionContext {
     userId?: string;
@@ -61,102 +62,6 @@ function generateRequestId(): string {
 }
 
 /**
- * Format attribute value for OTLP
- */
-function formatAttributeValue(value: unknown): Record<string, unknown> {
-    if (typeof value === 'string') {
-        return { stringValue: value };
-    }
-    if (typeof value === 'number') {
-        if (Number.isInteger(value)) {
-            return { intValue: String(value) };
-        }
-        return { doubleValue: value };
-    }
-    if (typeof value === 'boolean') {
-        return { boolValue: value };
-    }
-    if (Array.isArray(value)) {
-        return {
-            arrayValue: {
-                values: value.map(v => formatAttributeValue(v))
-            }
-        };
-    }
-    return { stringValue: String(value) };
-}
-
-/**
- * Convert severity text to OpenTelemetry severity number
- */
-function getSeverityNumber(severity: string): number {
-    const map: Record<string, number> = {
-        'trace': 1,
-        'debug': 5,
-        'info': 9,
-        'warn': 13,
-        'warning': 13,
-        'error': 17,
-        'fatal': 21,
-    };
-    return map[severity.toLowerCase()] || 9;
-}
-
-/**
- * Build OTLP payload for a single log entry
- */
-function buildOTLPPayload(
-    severityText: string,
-    message: string,
-    attributes: LogAttributes
-) {
-    const timestamp = new Date().toISOString();
-    const cleanAttributes: Record<string, string | number | boolean | string[] | number[]> = {};
-
-    for (const [key, value] of Object.entries(attributes)) {
-        if (value !== null && value !== undefined) {
-            cleanAttributes[key] = value as string | number | boolean | string[] | number[];
-        }
-    }
-
-    const enrichedAttributes = {
-        ...cleanAttributes,
-        'log.timestamp': timestamp,
-        'service.name': SERVICE_NAME,
-        'environment': process.env.NODE_ENV || 'development',
-    };
-
-    return {
-        resourceLogs: [{
-            resource: {
-                attributes: [
-                    { key: 'service.name', value: { stringValue: SERVICE_NAME } },
-                    { key: 'deployment.environment', value: { stringValue: process.env.NODE_ENV || 'development' } },
-                    { key: 'telemetry.sdk.name', value: { stringValue: 'posthog-logger' } },
-                    { key: 'telemetry.sdk.version', value: { stringValue: '1.0.0' } },
-                ]
-            },
-            scopeLogs: [{
-                scope: { name: SERVICE_NAME, version: '1.0.0' },
-                logRecords: [{
-                    timeUnixNano: String(new Date(timestamp).getTime() * 1000000),
-                    observedTimeUnixNano: String(Date.now() * 1000000),
-                    severityNumber: getSeverityNumber(severityText),
-                    severityText: severityText.toUpperCase(),
-                    body: { stringValue: message },
-                    attributes: Object.entries(enrichedAttributes)
-                        .filter(([, v]) => v !== null && v !== undefined)
-                        .map(([key, value]) => ({
-                            key,
-                            value: formatAttributeValue(value)
-                        }))
-                }]
-            }]
-        }]
-    };
-}
-
-/**
  * Send a log directly to PostHog (edge runtime compatible, no batching)
  * This is fire-and-forget - doesn't block the request
  */
@@ -171,8 +76,8 @@ function sendLogImmediate(
         return;
     }
 
-    const endpoint = `${POSTHOG_HOST.replace(/\/$/, '')}/i/v1/logs`;
-    const payload = buildOTLPPayload(severityText, message, attributes);
+    const endpoint = getLogsEndpoint();
+    const payload = buildOTLPPayloadSingle(severityText, message, attributes);
 
     // Fire-and-forget fetch - don't await, don't block
     fetch(endpoint, {
